requirements.txt
Flask>=2.2.0
Flask-Login>=0.6.0
Flask-WTF>=1.0.0
WTForms>=3.0.0
Flask-SQLAlchemy>=3.0.0
Werkzeug>=2.2.0
Jinja2>=3.1.0
itsdangerous>=2.1.0
click>=8.1.0
Pillow>=9.0.0
requests>=2.28.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.0
beautifulsoup4>=4.11.0 